+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## [0.11.2] (Unreleased)
 - Fix bug where generated code would not compile if a message was named Unit (#1126)
 - Add option to block generation/access to field getters (#1098)
-- 
+- Add option for enum annotation (#1155)
+- Add `unrecognized_value_class_is_private` option to make `Unrecognised` class private for safer development. (#1155)
+
 ## [0.11.1]
 - Support Scala 3.0.0-RC2
 - ScalaPB artifacts include [https://www.scala-lang.org/blog/2021/02/16/preventing-version-conflicts-with-versionscheme.html](versionscheme)